# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/candidate-accounts/me": {
      "get": {
        "operationId": "CandidateAccountController_findMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateAccount"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu hoặc token không hợp lệ."
          },
          "403": {
            "description": "Chỉ ứng viên mới có thể gọi endpoint này."
          },
          "404": {
            "description": "Không tìm thấy tài khoản ứng viên."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy thông tin tài khoản ứng viên",
        "tags": ["Candidate - Account"]
      },
      "patch": {
        "operationId": "CandidateAccountController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMyCandidateAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateAccount"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu yêu cầu không hợp lệ."
          },
          "401": {
            "description": "Thiếu hoặc token không hợp lệ."
          },
          "403": {
            "description": "Chỉ ứng viên mới có thể gọi endpoint này."
          },
          "404": {
            "description": "Không tìm thấy tài khoản ứng viên."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ứng viên đổi mật khẩu tài khoản",
        "tags": ["Candidate - Account"]
      }
    },
    "/candidate-accounts": {
      "get": {
        "operationId": "CandidateAccountController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateAccountList"
                }
              }
            }
          },
          "400": {
            "description": "Tham số truy vấn không hợp lệ."
          },
          "401": {
            "description": "Thiếu hoặc token không hợp lệ."
          },
          "403": {
            "description": "Chỉ admin mới có thể gọi endpoint này."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách tài khoản ứng viên (ADMIN)",
        "tags": ["Candidate - Account"]
      }
    },
    "/candidate-accounts/{id}": {
      "get": {
        "operationId": "CandidateAccountController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "1f5f4a65-50d7-4f24-a65f-4f2a4d42f9cf",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateAccount"
                }
              }
            }
          },
          "401": {
            "description": "Thiếu hoặc token không hợp lệ."
          },
          "403": {
            "description": "Chỉ admin mới có thể gọi endpoint này."
          },
          "404": {
            "description": "Không tìm thấy tài khoản ứng viên."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy thông tin chi tiết tài khoản ứng viên (ADMIN)",
        "tags": ["Candidate - Account"]
      }
    },
    "/candidate-accounts/{id}/status": {
      "patch": {
        "operationId": "CandidateAccountController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "1f5f4a65-50d7-4f24-a65f-4f2a4d42f9cf",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCandidateAccountStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateAccount"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu yêu cầu không hợp lệ."
          },
          "401": {
            "description": "Thiếu hoặc token không hợp lệ."
          },
          "403": {
            "description": "Chỉ admin mới có thể gọi endpoint này."
          },
          "404": {
            "description": "Không tìm thấy tài khoản ứng viên."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admin cập nhật trạng thái tài khoản ứng viên",
        "tags": ["Candidate - Account"]
      }
    },
...
    "/candidate-profiles/me/languages": {
      "get": {
        "operationId": "CandidateLanguagesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateLanguage"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách ngoại ngữ",
        "tags": ["Candidate - Languages"]
      },
      "post": {
        "operationId": "CandidateLanguagesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidateLanguageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateLanguage"
                }
              }
            }
          },
          "409": {
            "description": "Language already exists in this profile."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create my language",
        "tags": ["Candidate - Languages"]
      }
    },
    "/candidate-profiles/me/languages/{id}": {
      "patch": {
        "operationId": "CandidateLanguagesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCandidateLanguageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateLanguage"
                }
              }
            }
          },
          "409": {
            "description": "Language already exists in this profile."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update my language",
        "tags": ["Candidate - Languages"]
      },
      "delete": {
        "operationId": "CandidateLanguagesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete my language",
        "tags": ["Candidate - Languages"]
      }
    },
...
  },
  "info": {
    "title": "UpNext API",
    "description": "Backend API for the UpNext IT recruitment platform.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CandidateAccount": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx4q8z1j0000u8p4e1o9v6m2"
          },
          "fullName": {
            "type": "string",
            "example": "Phan Đức Toàn"
          },
          "email": {
            "type": "string",
            "example": "pductoandev@gmail.com"
          },
          "authProvider": {
            "type": "string",
            "enum": ["DEFAULT", "GOOGLE", "GITHUB", "LINKEDIN"],
            "example": "DEFAULT"
          },
          "providerUserId": {
            "type": "object",
            "example": "google-oauth2|1234567890",
            "nullable": true
          },
          "candidateAccountStatus": {
            "type": "string",
            "enum": ["ACTIVE", "BANNED", "PENDING_VERIFICATION"],
            "example": "ACTIVE"
          },
          "emailVerifiedAt": {
            "type": "object",
            "example": "2026-06-09T04:30:00.000Z",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-09T04:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-09T04:30:00.000Z"
          }
        },
        "required": [
          "id",
          "fullName",
          "email",
          "authProvider",
          "candidateAccountStatus",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateMyCandidateAccountDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72,
            "writeOnly": true
          }
        }
      },
      "CandidateAccountListMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 42
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": ["page", "limit", "total", "totalPages"]
      },
      "CandidateAccountList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateAccount"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/CandidateAccountListMeta"
          }
        },
        "required": ["items", "meta"]
      },
      "UpdateCandidateAccountStatusDto": {
        "type": "object",
        "properties": {
          "candidateAccountStatus": {
            "type": "string",
            "enum": ["ACTIVE", "BANNED", "PENDING_VERIFICATION"],
            "example": "BANNED"
          }
        },
        "required": ["candidateAccountStatus"]
      },
...
      "CandidateLanguage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "candidateProfileId": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "proficiency": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "candidateProfileId",
          "language",
          "proficiency",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateCandidateLanguageDto": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "example": "English",
            "maxLength": 80
          },
          "proficiency": {
            "type": "string",
            "example": "Intermediate",
            "maxLength": 80
          }
        },
        "required": ["language", "proficiency"]
      },
      "UpdateCandidateLanguageDto": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "example": "English",
            "maxLength": 80
          },
          "proficiency": {
            "type": "string",
            "example": "Intermediate",
            "maxLength": 80
          }
        }
      },
...
    }
  }
}